refactor(receiver-mock): move handlers for querying metrics data to a separate module
